backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
google-generativeai==0.8.5
python-dotenv==1.1.1 

